model_api/requirements.txt
fastapi
uvicorn
tensorflow
pandas
numpy
minio
python-dotenv
